# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "tags": [
          "App"
        ]
      }
    },
    "/users": {
      "get": {
        "operationId": "UserController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "users"
        ]
      },
      "post": {
        "operationId": "UserController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UserController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "UserController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UserController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "users"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/images": {
      "post": {
        "operationId": "ImageController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateImageRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateImageResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "images"
        ]
      },
      "get": {
        "operationId": "ImageController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ImageResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "images"
        ]
      }
    },
    "/images/{id}": {
      "get": {
        "operationId": "ImageController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "images"
        ]
      },
      "delete": {
        "operationId": "ImageController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "images"
        ]
      }
    },
    "/images/{id}/download-url": {
      "get": {
        "operationId": "ImageController_getDownloadUrl",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadUrlResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "images"
        ]
      }
    }
  },
  "info": {
    "title": "Petal API",
    "description": "Petal 画像コンテンツ管理 API",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Cognito アクセストークン"
      }
    },
    "schemas": {
      "UserRole": {
        "type": "string",
        "enum": [
          "admin",
          "user"
        ]
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "role": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UserRole"
              }
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "deletedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "id": {
            "type": "string"
          },
          "cognitoSub": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "nameKana": {
            "type": "string"
          }
        },
        "required": [
          "role",
          "createdAt",
          "updatedAt",
          "deletedAt",
          "id",
          "cognitoSub",
          "name",
          "nameKana"
        ]
      },
      "CreateUserRequestDto": {
        "type": "object",
        "properties": {
          "role": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UserRole"
              }
            ]
          },
          "cognitoSub": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "nameKana": {
            "type": "string"
          }
        },
        "required": [
          "cognitoSub",
          "name",
          "nameKana"
        ]
      },
      "UpdateUserRequestDto": {
        "type": "object",
        "properties": {
          "role": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UserRole"
              }
            ]
          },
          "name": {
            "type": "string"
          },
          "nameKana": {
            "type": "string"
          }
        }
      },
      "LoginRequestDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "idToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "expiresIn": {
            "type": "number"
          },
          "email": {
            "type": "string"
          }
        },
        "required": [
          "accessToken",
          "idToken",
          "refreshToken",
          "expiresIn",
          "email"
        ]
      },
      "CreateImageRequestDto": {
        "type": "object",
        "properties": {
          "mimeType": {
            "type": "string",
            "enum": [
              "image/jpeg",
              "image/png",
              "image/gif",
              "image/webp"
            ]
          },
          "originalFilename": {
            "type": "string"
          },
          "sizeBytes": {
            "type": "number"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "mimeType",
          "originalFilename",
          "sizeBytes"
        ]
      },
      "ImageResponseDto": {
        "type": "object",
        "properties": {
          "mimeType": {
            "type": "string",
            "enum": [
              "image/jpeg",
              "image/png",
              "image/gif",
              "image/webp"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string"
          },
          "ownerUserId": {
            "type": "string"
          },
          "originalFilename": {
            "type": "string"
          },
          "sizeBytes": {
            "type": "number"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "mimeType",
          "createdAt",
          "updatedAt",
          "id",
          "ownerUserId",
          "originalFilename",
          "sizeBytes",
          "title",
          "description"
        ]
      },
      "UploadInstructionHeadersDto": {
        "type": "object",
        "properties": {
          "Content-Type": {
            "type": "string"
          }
        },
        "required": [
          "Content-Type"
        ]
      },
      "UploadInstructionDto": {
        "type": "object",
        "properties": {
          "method": {
            "type": "string",
            "enum": [
              "PUT"
            ]
          },
          "url": {
            "type": "string"
          },
          "expiresInSeconds": {
            "type": "number"
          },
          "headers": {
            "$ref": "#/components/schemas/UploadInstructionHeadersDto"
          }
        },
        "required": [
          "method",
          "url",
          "expiresInSeconds",
          "headers"
        ]
      },
      "CreateImageResponseDto": {
        "type": "object",
        "properties": {
          "image": {
            "$ref": "#/components/schemas/ImageResponseDto"
          },
          "upload": {
            "$ref": "#/components/schemas/UploadInstructionDto"
          }
        },
        "required": [
          "image",
          "upload"
        ]
      },
      "DownloadUrlResponseDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          },
          "expiresInSeconds": {
            "type": "number"
          }
        },
        "required": [
          "url",
          "expiresInSeconds"
        ]
      }
    }
  }
}